Add tests for promise branching
